Remove old route files.
