--- client/src/routes/Routes.paths.ts
@@ -1,28 +1,0 @@
-export enum RoutingPath {
-  ROOT = '/',
-  // LOGIN = '/login',
-  // STUDENTOVERVIEW = '/studentoverview',
-  // STUDENT_INFO = '/studentinfo/student/:studentId/:tutorialId?',
-  // TEAMOVERVIEW = '/teamoverview',
-  // ATTENDANCE = '/attendance',
-  // ENTER_POINTS_OVERVIEW = '/enterpoints/:sheetId?',
-  // ENTER_POINTS_TEAM = '/enterpoints/:sheetId/team/:teamId',
-  // ENTER_POINTS_STUDENT = '/enterpoints/:sheetId/team/:teamId/student/:studentId',
-  // PRESENTATION_POINTS = '/presentations/:sheetId?',
-  // SCHEIN_EXAMS_OVERVIEW = '/scheinexams/:examId?',
-  // SCHEIN_EXAMS_STUDENT = '/scheinexams/:examId/student/:studentId',
-  // DASHBOARD = '/dashboard',
-  // MANAGE_USERS = '/admin/usermanagement',
-  // IMPORT_USERS = '/admin/usermanagement/generate',
-  // MANAGE_TUTORIALS = '/admin/tutorialmanagement',
-  // MANAGE_TUTORIAL_INTERNALS = '/admin/tutorialmanagement/manage/:tutorialId',
-  // GENERATE_TUTORIALS = '/admin/tutorialmanagement/generate',
-  // MANAGE_TUTORIALS_SUBSTITUTES = '/admin/tutorialmanagement/substitutes/:tutorialId/substitute',
-  // MANAGE_SCHEIN_CRITERIAS = '/admin/scheincriterias',
-  // SCHEIN_CRITERIAS_INFO = '/admin/scheincriterias/info/:id',
-  // MANAGE_ATTENDANCES = '/admin/attendances',
-  // MANAGE_SHEETS = '/admin/sheets',
-  // MANAGE_ALL_STUDENTS = '/admin/students',
-  // MANAGE_SCHEIN_EXAMS = '/admin/scheinexams',
-  // IMPORT_TUTORIALS_AND_USERS = '/admin/import',
-}
